refactor: move sortTeamPlayersFirst to lib/teamPlayers; drop unused import
